+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add Media & Text (Dude) block variation with placeholder image and default content, Ref: DEV-245
 * Add `wp_localize_script` to pass theme URL to block editor JS, Ref: DEV-245
 * Remove `_editor.scss` ACF-specific editor styles, Ref: DEV-245
+* Fix block editor styles for iframed editor by using `enqueue_block_assets` hook, Ref: DEV-245
 
 ### 9.7.0: 2026-01-09